+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - extract mandatory fields for igs variables and variable groups
 - update mex-common dependency to 1.4
 - create s3 sink base class and bequeath to s3Sink (writes ndjson) and new S3XlsxSink (writs xlsx)
+- publisher: replace references to non-consenting person ids with their respective unit ids
 
 ### Changes